Files for the basic resume website
These are the files I'd like to use across machines to design the 2-button (an anchor and a button) website that loads my resume. I use an anchor so users can use the middle mouse button to open the resume. I aim to learn how to add functionality to button elements so that users can use the middle mouse button on website buttons just like they would an anchor.